# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tug API",
        "description": "API for managing vehicles and drivers in a logistics system.",
        "version": "1.0"
    },
    "paths": {
        "/api/drivers": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Retrieve a list of all drivers",
                "operationId": "90f1e2d7ba23baa1e136288ed235666c",
                "responses": {
                    "200": {
                        "description": "List of drivers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DriverResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Create a new driver",
                "description": "Create a new driver.",
                "operationId": "6bdf167696d642223e0d27e88a0b4896",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "license_number",
                                    "phone_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "license_number": {
                                        "type": "string",
                                        "example": "ABC12345"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Driver created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriverResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Retrieve a list of all vehicles",
                "operationId": "727634867f59796290a2fff276979198",
                "responses": {
                    "200": {
                        "description": "List of vehicles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/VehicleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Create a new vehicle",
                "description": "Create a new vehicle.",
                "operationId": "a88e0bca4193c91e713d0441659d199c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plate_number",
                                    "brand",
                                    "model",
                                    "year"
                                ],
                                "properties": {
                                    "plate_number": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Corolla"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2022
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vehicles/{vehicle}/assign-driver": {
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Assign a driver to a vehicle",
                "description": "Assign a driver to a vehicle.",
                "operationId": "9f38811029ce36d49323baa76cd98f97",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "ID of the vehicle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "driver_id"
                                ],
                                "properties": {
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Driver assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to assign driver",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DriverResource": {
                "title": "Driver Resource",
                "description": "Driver model representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "license_number": {
                        "type": "string",
                        "example": "ABC12345"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "vehicles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VehicleResource"
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResource": {
                "title": "Error Response",
                "description": "Standard error response format",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": "Error message"
                        }
                    }
                },
                "type": "object"
            },
            "VehicleResource": {
                "title": "Vehicle Resource",
                "description": "Vehicle model representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plate_number": {
                        "type": "string",
                        "example": "ABC123"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2022
                    },
                    "driver": {
                        "$ref": "#/components/schemas/DriverResource"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Drivers",
            "description": "API Endpoints for managing drivers"
        },
        {
            "name": "Vehicles",
            "description": "API Endpoints for managing vehicles"
        }
    ]
}